fix: free memory used by proposals in firewood
Problem: Previously, when the database is closing, coreth would not free
the memory used by the proposals in firewood. This is relying on an
internal implementation detail of firewood but it is a violation of the
firewood API contract. In the [future], this will lead to memory leaks.

Solution: This change adds a method to the ProposalContext that the db
uses to recursively free the proposal but skip the proposal map removal
as it is redundant when closing the database.

[future]: https://github.com/ava-labs/firewood/pull/1221