Wait for tx pool event loop
Currently whenever we ask the VM to wait for an event, it checks if the mempool has transactions in it, and if so, it proceeds to return a "pending transactions" event.

The mempool has or has not transactions in it according to the block that the VM is configured to build on top on.

When the VM is being told to change the preferred block to build the next block on, the mempool is asynchronously re-organized in the background according to the block preference change.

Since the mempool re-organization happens asynchronously and in the background, when a VM changes its preference and is immediately asked to build a new block, the mempool may still contain transactions in it, because the mempool re-organization is still being performed asynchronously in the background. This leads to a false positive, as the mempool may not have transactions in it after its re-organization, but the WaitForEvent API call may return that the mempool has transactions in it and would result in building a block that shouldn't have been built.

This PR fixes this false positive, as explained below:

When we set the preference in the VM or accept a block, we mark the block height to be pending.
In parallel, we subscribe to updates from the mempool which are sent only once the re-org corresponding to a block has finished. Upon such an update, we mark the block as the latest block.

When the VM checks whether to build a block, additionally to checking whether there are transactions in the mempool, we now also check whether the mempool is pending a re-org. A mempool is pending a re-org if the block number that is set when the block is accepted or the preference changes (the pending number) is different than the block number that is received in the latest subscription update from the mempool (the latest number).